requirements.txt
altair==5.1.2
appdirs==1.4.4
atomicwrites==1.4.0
attrs==23.1.0
backports.weakref==1.0.post1
blinker==1.7.0
branca==0.6.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==15.1.0
contourpy==1.2.0
cssselect==1.1.0
cycler==0.12.1
daal==2023.1.1
daal4py==2023.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
finance-datareader==0.9.50
folium==0.14.0
fonttools==4.25.0
fst-pso==1.8.1
funcy==2.0
FuzzyTM==2.0.5
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
jsonpointer==2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
konlpy==0.6.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
miniful==0.0.6
munkres==1.1.4
mypy-extensions==0.4.3
numexpr==2.8.7
numpy==1.26.2
openpyxl==3.0.10
outcome==1.2.0
packaging==23.2
pandas==2.1.3
patsy==0.5.3
pep8==1.7.1
Pillow==9.4.0
ply==3.11
protobuf==4.25.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
pyFUME==0.2.25
Pygments==2.17.0
pyLDAvis==3.4.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
rpds-py==0.13.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
selenium==4.12.0
simpful==2.11.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
statsmodels==0.14.0
streamlit==1.28.1
tbb==2021.11.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
wordcloud==1.9.2
wsproto==1.2.0
zipp==3.17.0
zstandard==0.19.0
